Add more consistency with the real code
