+++ types/lasso/lib/writers/index.d.ts
@@ -1,0 +1,8 @@
+import { LassoConfig } from '../Lasso';
+import _Writer from './Writer';
+
+export interface Writer extends _Writer {}
+
+export function fileWriter(fileWriterConfig: any, lassoConfig: LassoConfig): any;
+
+export function createWriter(writerImpl: any): Writer;
