+++ types/lasso/middleware/index.d.ts
@@ -1,0 +1,4 @@
+import { RequestHandler } from 'express';
+import Lasso from '../lib/Lasso';
+
+export function serveStatic(options?: { lasso?: Lasso; sendOptions?: any; }): RequestHandler;
